# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API React",
        "description": "Descrição da API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "summary": "Get Users",
                "description": "Retrieve a paginated list of users.",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new user",
                "description": "Create a new user with specified details.",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "summary": "Show User",
                "description": "Retrieve a single user by ID.",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário retornado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update User",
                "description": "Update user details by ID.",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First name of the user",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role_id": {
                                        "description": "ID of the user's role",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados enviados."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete User",
                "description": "Delete a user by ID.",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "summary": "Get authenticated user",
                "description": "Retrieve the currently authenticated user.",
                "operationId": "fd8e4e27fd1a9e98a6638cec7e2126ca",
                "responses": {
                    "200": {
                        "description": "Usuário autenticado retornado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/info": {
            "put": {
                "summary": "Update user information",
                "description": "Update the authenticated user's information.",
                "operationId": "96af1e0d1f720f5395533b025cbcb9b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First name of the user",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role_id": {
                                        "description": "ID of the user's role",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Informações do usuário atualizadas com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/password": {
            "put": {
                "summary": "Update user password",
                "description": "Update the authenticated user's password.",
                "operationId": "f0598ccf9b10c1c1e00f532bbcad0fba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "New password for the user",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Senha do usuário atualizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserCreateRequest": {
                "required": [
                    "last_name",
                    "email",
                    "first_name",
                    "role_id",
                    "password"
                ],
                "properties": {
                    "first_name": {
                        "description": "First name of the user",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "role_id": {
                        "description": "ID of the user's role",
                        "type": "integer"
                    },
                    "password": {
                        "description": "Password for the user account",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer"
                    },
                    "first_name": {
                        "description": "Primeiro nome do usuário",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Último nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "Função ou cargo do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação por JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}